+++ frontend/src/types/Page.ts
@@ -1,0 +1,5 @@
+export type PageInfo = {
+  title: string;
+  url: string;
+  content: string;
+};
